login编写完成但是无法访问,排查错误

+++ core/model/user_basic.go
@@ -1,7 +1,7 @@
 package model
 
 type UserBasic struct {
-	Id       int
+	Id       uint64
 	Identity string
 	Name     string
 	Password string
